#33 Refactor: Adjust to PR comments
