+++ packages/ui/src/components/Textarea/index.ts
@@ -1,1 +1,2 @@
-export * from './Textarea';
+export { Textarea } from './Textarea';
+export type { TextareaProps } from './Textarea';
